Rename files to use UpperCamelCaseCluster/Logic for boolean-state-configuration-server
