+++ src/app/clusters/boolean-state-configuration-server/BUILD.gn
@@ -16,8 +16,8 @@ import("//build_overrides/chip.gni")
 
 source_set("boolean-state-configuration-server") {
   sources = [
-    "boolean-state-configuration-cluster.cpp",
-    "boolean-state-configuration-cluster.h",
+    "BooleanStateConfigurationCluster.cpp",
+    "BooleanStateConfigurationCluster.h",
     "boolean-state-configuration-delegate.h",
   ]
